+++ src/main/java/io/github/mortuusars/sootychimneys/SootyChimneys.java
@@ -1,7 +1,7 @@
 package io.github.mortuusars.sootychimneys;
 
 import io.github.mortuusars.sootychimneys.integration.create.CreateIntegration;
-import io.github.mortuusars.sootychimneys.config.CommonConfig;
+import io.github.mortuusars.sootychimneys.config.Config;
 import io.github.mortuusars.sootychimneys.core.WindGetter;
 import io.github.mortuusars.sootychimneys.setup.Registry;
 import net.minecraft.resources.ResourceLocation;
@@ -19,22 +19,20 @@ import java.util.Random;
 public class SootyChimneys
 {
     public static final String MOD_ID = "sootychimneys";
-
     public static final Random RANDOM = new Random();
-
-    public static boolean debug = false;
 
     public SootyChimneys()
     {
-        ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, CommonConfig.SPEC);
+        ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, Config.SPEC);
 
         Registry.init();
         MinecraftForge.EVENT_BUS.addListener(WindGetter::onPlayerTick);
 
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::enqueueIMC);
+    }
 
-        if (debug)
-            MinecraftForge.EVENT_BUS.addListener(WindGetter::onRightClick);
+    public static ResourceLocation resource(String path) {
+        return new ResourceLocation(MOD_ID, path);
     }
 
     private void enqueueIMC(final InterModEnqueueEvent event)
@@ -42,9 +40,5 @@ public class SootyChimneys
         if (ModList.get().isLoaded("create")) {
             CreateIntegration.registerMovingBehaviors();
         }
-    }
-
-    public static ResourceLocation resource(String path) {
-        return new ResourceLocation(MOD_ID, path);
     }
 }