+++ src/main/java/io/github/mortuusars/sootychimneys/block/BrickChimneyBlock.java
@@ -16,11 +16,12 @@ import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
 
-@SuppressWarnings({"deprecation", "NullableProblems"})
 public class BrickChimneyBlock extends ChimneyBlock implements ISootyChimney {
-    private static final VoxelShape _shape = Shapes.or(
+    private static final VoxelShape SHAPE = Shapes.or(
             Block.box(1d, 0d, 1d, 15d,11d,15d),
             Block.box(0d, 11d, 0d, 16d,16d,16d));
+
+    private static final VoxelShape STACKED_SHAPE = Block.box(1d, 0d, 1d, 15d, 16d, 15d);
 
     public BrickChimneyBlock() {
         super(new ChimneySmokeProperties(0.5f, 1f, 0.5f, 0.25f, 0.1f, 0.25f)
@@ -33,14 +34,19 @@ public class BrickChimneyBlock extends ChimneyBlock implements ISootyChimney {
     }
 
     @Override
-    public @NotNull VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext) {
-        return _shape;
+    protected VoxelShape getDefaultShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
+        return SHAPE;
+    }
+
+    @Override
+    protected VoxelShape getStackedShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
+        return STACKED_SHAPE;
     }
 
     @NotNull
     @Override
     public Block getCleanVariant() {
-        return ModBlocks.BRICK_CHIMNEY.get(); //can't return 'this' here because Dirty variant will have wrong clean block, and 'this' breaks scraping with axe
+        return ModBlocks.BRICK_CHIMNEY.get();
     }
 
     @NotNull