+++ src/main/java/io/github/mortuusars/sootychimneys/block/CopperChimneyBlock.java
@@ -14,15 +14,14 @@ import net.minecraft.world.level.material.MaterialColor;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
-import org.jetbrains.annotations.NotNull;
 
-@SuppressWarnings("NullableProblems")
 public class CopperChimneyBlock extends ChimneyBlock implements ISootyChimney {
-
-    private static final VoxelShape _shape = Shapes.or(
+    private static final VoxelShape SHAPE = Shapes.or(
             Block.box(5,0,5, 11,4,11),
             Block.box(6,4,6, 10,16,10),
             Block.box(5,10,5, 11,14,11));
+
+    private static final VoxelShape STACKED_SHAPE = Block.box(5, 0, 5, 11, 16, 11);
 
     public CopperChimneyBlock() {
         super(new ChimneySmokeProperties(0.5f, 1.25f, 0.5f, 0.025f, 0.05f, 0.025f)
@@ -35,10 +34,14 @@ public class CopperChimneyBlock extends ChimneyBlock implements ISootyChimney {
                     .requiresCorrectToolForDrops());
     }
 
-    @SuppressWarnings("deprecation")
     @Override
-    public @NotNull VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext) {
-        return _shape;
+    protected VoxelShape getDefaultShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
+        return SHAPE;
+    }
+
+    @Override
+    protected VoxelShape getStackedShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
+        return STACKED_SHAPE;
     }
 
     @Override