+++ src/main/java/io/github/mortuusars/sootychimneys/block/StoneBrickChimneyBlock.java
@@ -14,13 +14,13 @@ import net.minecraft.world.level.material.MaterialColor;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
-import org.jetbrains.annotations.NotNull;
 
-@SuppressWarnings({"deprecation", "NullableProblems"})
 public class StoneBrickChimneyBlock extends ChimneyBlock implements ISootyChimney {
-    private static final VoxelShape  _shape = Shapes.or(
+    private static final VoxelShape SHAPE = Shapes.or(
             Block.box(4,0,4, 12,11,12),
             Block.box(3, 11, 3, 13, 16, 13));
+
+    private static final VoxelShape STACKED_SHAPE = Block.box(4,0,4, 12,16,12);
 
     public StoneBrickChimneyBlock() {
         super(new ChimneySmokeProperties(0.5f, 1.2f, 0.5f, 0.025f, 0.05f, 0.025f)
@@ -33,8 +33,13 @@ public class StoneBrickChimneyBlock extends ChimneyBlock implements ISootyChimne
     }
 
     @Override
-    public @NotNull VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext) {
-        return _shape;
+    protected VoxelShape getDefaultShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
+        return SHAPE;
+    }
+
+    @Override
+    protected VoxelShape getStackedShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
+        return STACKED_SHAPE;
     }
 
     @Override