+++ src/main/java/io/github/mortuusars/sootychimneys/setup/ModTags.java
@@ -1,5 +1,6 @@
 package io.github.mortuusars.sootychimneys.setup;
 
+import io.github.mortuusars.sootychimneys.SootyChimneys;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.ItemTags;
@@ -14,5 +15,6 @@ public class ModTags {
 
     public static class Blocks {
         public static final TagKey<Block> CHIMNEYS = BlockTags.create(new ResourceLocation("forge:chimneys"));
+        public static final TagKey<Block> SMOKE_BLOCKING = BlockTags.create(SootyChimneys.resource("smoke_blocking"));
     }
 }